requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.3
decorator==4.4.0
dj-database-url==0.5.0
Django==2.1.2
django-bootstrap4==0.0.8
django-debug-toolbar==1.11
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.4
ez-setup==0.9
flake8==3.6.0
gunicorn==19.9.0
idna==2.6
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.1
mccabe==0.6.1
more-itertools==7.0.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.11.0
prompt-toolkit==2.0.9
psycopg2==2.7.5
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.4.0
pytest==4.5.0
python-decouple==3.1
pytz==2018.4
requests==2.18.4
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==4.1
